requirements.txt
python-telegram-bot==20.3
pyswisseph
python-dotenv
aiohttp
flask